desktop_app/requirements.txt
PyQt5
pdfplumber
scikit-learn==1.8.0
joblib
openai
python-dotenv
